+++ src/h3lib/lib/h3Index.c
@@ -33,11 +33,6 @@
 #include "mathExtensions.h"
 #include "vertex.h"
 
-// TODO: https://github.com/uber/h3/issues/984
-static const bool isBaseCellPentagonArr[128] = {
-    [4] = 1,  [14] = 1, [24] = 1, [38] = 1, [49] = 1,  [58] = 1,
-    [63] = 1, [72] = 1, [83] = 1, [97] = 1, [107] = 1, [117] = 1};
-
 /** @var H3ErrorDescriptions
  *  @brief An array of strings describing each of the H3ErrorCodes enum values
  */
@@ -146,7 +141,7 @@ H3Error H3_EXPORT(constructCell)(int res, int baseCellNumber, const int *digits,
     H3_SET_RESOLUTION(h, res);
     H3_SET_BASE_CELL(h, baseCellNumber);
 
-    bool isPentagon = isBaseCellPentagonArr[baseCellNumber];
+    bool isPentagon = baseCellIsPentagon[baseCellNumber];
 
     for (int r = 1; r <= res; r++) {
         int d = digits[r - 1];
@@ -325,7 +320,7 @@ We can check that (in the lower 45 = 15*3 bits) the position of the
 first 1 bit isn't divisible by 3.
 */
 static inline bool _hasDeletedSubsequence(H3Index h, int base_cell) {
-    if (isBaseCellPentagonArr[base_cell]) {
+    if (baseCellIsPentagon[base_cell]) {
         h <<= 19;
         h >>= 19;